Refactored the QuestionCreationStrategy package for math questions. Created a randomBean and injected it into all classes that use the Random() object. Added tests for all QuestionService classes. Added tests for QuestionServiceFactoryImp Added tests for RandomUtils Added tests for UserServiceImp

+++ src/main/java/org/skypro/be/quizmaster/model/Result.java
@@ -23,15 +23,15 @@ public class Result {
     private LocalDateTime examPassed;
 
     @Column
-    private long totalQuestions;
+    private int totalQuestions;
 
     @Column
-    private long correctAnswers;
+    private int correctAnswers;
 
     public Result() {
     }
 
-    public Result(Long userId, String description, LocalDateTime examPassed, long totalQuestions, long correctAnswers) {
+    public Result(Long userId, String description, LocalDateTime examPassed, int totalQuestions, int correctAnswers) {
         this.userId = userId;
         this.description = description;
         this.examPassed = examPassed;
